+++ NEWS.md
@@ -1,3 +1,13 @@
+# did 2.1.2
+
+  * Added wrapper function for HonestDiD package
+  
+  * Fix bug for setups where `gname` is not contained in `tname` (but is in the `tname` range)
+  
+  * Fix bug for including too many groups with universal base period in pre-treatment periods
+  
+  * Bug fix for anticipation using `notyettreated` comparison group
+
 # did 2.1.1
 
   * Bug fixes related unbalanced panel and clustered standard errors